% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Plotting Class ryx}
\usage{
\method{plot}{ryx}(results, ...)
}
\arguments{
\item{results}{results from ryx}
}
\value{
A barbell plot of absolute value of the correlation values, with color representing the direction of the correlation.
}
\description{
Plots correlations between a dependent variable y and one or more independent x variables using output from ryx function from ryx package.
}
\examples{
x <- ryx(mtcars, y="mpg", x=c("hp", "wt", "disp", "cyl", "am", "gear"))
plot(x)
}
