% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{RYX: Correlation Table}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{a data frame}

\item{y}{the name of a numeric variable in the data frame}

\item{x}{a character vector with the names of one or more numeric variables in the data frame}

\item{digits}{an integer. Indicates the number of decimal places to print (3 is the default)}
}
\value{
A list with values $y, dependent variable, $x, dependent variables, $df, a data frame with correlation values; r values, p-values and significance levels, of class "ryx".
}
\description{
Prints a table of correlations between a dependent variable y and one or more independent x variables. Uses only base R functions like cor.test(), for(), and rbind().
}
\examples{
ryx(mtcars, y="mpg", x=c("hp", "wt", "disp", "cyl", "am", "gear"))
ryx(mtcars, y="mpg", x=c("hp", "wt", "disp"))
}
