% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Gives a summary of results from ryx function.}
\usage{
\method{summary}{ryx}(results, ...)
}
\arguments{
\item{results}{results from ryx()}
}
\value{
A string of summarized results from ryx, including variables used, median absolute correlation value and range, and how many significant variables were found in the analysis.
}
\description{
Prints a summary of results from ryx function detailing the median absolute correlation value with ranges, as well as how many variables were significant at the p < 0.05 level.
}
\examples{
x <- ryx(mtcars, y="mpg", x=c("hp", "wt", "disp", "cyl", "am", "gear"))
summary(x)
}
