% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx_methods.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Plot method for ryx objects}
\usage{
\method{plot}{ryx}(x, ...)
}
\arguments{
\item{x}{An object of class ryx.}
}
\description{
Plot method for ryx objects
}
\examples{
library(MASS)
x <- ryx(Boston, y = "medv")
plot(x)

}
