% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Correlation Table}
\usage{
\method{plot}{ryx}(x, ...)
}
\arguments{
\item{x}{object inherited from ryx function}

\item{...}{(optional), any additional parameters}

\item{data}{a data frame}
}
\value{
method used, dataset used, variables used for outliers detected, indices of any detected outliers, scores for the outliers, and values for optional parameters
}
\description{
Creates a correlation table using numeric variables.
}
\examples{
library(MASS)
x <- ryx(Boston, "medv")
plot(x)
}
