% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Correlation Table Graph}
\usage{
\method{plot}{ryx}(x, ...)
}
\arguments{
\item{x}{object inherited from ryx function}

\item{...}{(optional), any additional parameters}
}
\value{
graphical representation of correlation
}
\description{
Creates a correlation table using numeric variables.
}
\examples{
library(MASS)
x <- ryx(Boston, "medv")
plot(x)
}
