% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Computes correlation of a data frame}
\usage{
\method{plot}{ryx}(x)
}
\arguments{
\item{x}{an object of class \code{ryx}}
}
\value{
a line segment ggplot
}
\description{
\code{plot.ryx} plots magnitude and direction of each correlation.
}
\examples{
\dontrun{
x <- ryx(MASS::Boston, y = "medv")
plot(x)
}
}
