% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Plot of Correlation for Object Class 'ryx'}
\usage{
\method{plot}{ryx}(x)
}
\arguments{
\item{x}{An object of 'ryx' class}
}
\value{
A plot with x-axis as the absolute correlation values and y-axis as the variables.
}
\description{
Given an object of class 'ryx', the \code{plot} function would return a ggplot showing the
absolute correlation values for each variables in the object.
}
\examples{
\dontrun{
#import MASS package for example data
library(MASS)

#create an 'ryx' class object
x <- ryx(Boston, y = "medv")

#plot the 'ryx' class object
plotR <- plot(x)
}

}
