% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Plot function for \code{ryx} objects}
\usage{
\method{plot}{ryx}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ryx}, which is the result of the \code{ryx} function.}
}
\value{
This function returns a \code{ggplot} object that visualizes the following:
\itemize{
\item Grey lines indicating the strength of correlation for each variable.
\item Colored dots at the end of each line to indicate the correlation direction (red for negative, blue for positive).
\item The x-axis represents the absolute value of the correlation coefficient, ranging from 0 to 1.
\item The y-axis shows the names of the variables.
}
}
\description{
This function creates a plot to visualize the correlations between the dependent variable (\code{y}) and independent variables (\code{x}). The plot shows a series of horizontal lines representing the correlations and colored dots at the end of each line to indicate the direction of the correlation: red for negative and blue for positive. The x-axis represents the absolute value of the correlation coefficient, and the y-axis represents the variables.
}
\examples{
x <- ryx(Boston, y = "medv")
plot(x)
}
