% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxy.R
\name{plot.ryx}
\alias{plot.ryx}
\title{plot.ryx}
\usage{
\method{plot}{ryx}(x)
}
\description{
The plot function takes correlations
created by the ryx function and prints
a plot of r values
}
\examples{
library(MASS)
x <- ryx(Boston, y="medv")
plot(x)

}
