% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Visualize \code{ryx} object}
\usage{
\method{plot}{ryx}(x, ...)
}
\arguments{
\item{x}{A \code{ryx} object}

\item{...}{Optional parameters passed to import functions}
}
\description{
Visualize the correlations of the given response and explanatory variables.
}
\examples{
x=ryx(mtcars, 'mpg', c('wt', 'hp'))
plot(x)
}
