% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ryx.R
\name{plot.ryx}
\alias{plot.ryx}
\title{plot.ryx}
\usage{
\method{plot}{ryx}(x)
}
\description{
This \code{plot.ryx} function can plot the ryx objects.

This is wonderful.
}
\examples{
\dontrun{
plot.ryx(myryx)
}
}
