% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx_methods.R
\name{print.ryx}
\alias{print.ryx}
\title{Print method for ryx objects}
\usage{
\method{print}{ryx}(x, ...)
}
\arguments{
\item{x}{An object of class ryx.}
}
\description{
Print method for ryx objects
}
\examples{
library(MASS)
x <- ryx(Boston, y = "medv")
print(x)
}
