% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ryx.R, R/print.ryx.R
\name{print.ryx}
\alias{print.ryx}
\title{Correlation Table}
\usage{
\method{print}{ryx}(x, ...)

\method{print}{ryx}(x, ...)
}
\arguments{
\item{x}{string, a list of numeric variables. Defaults to all variables in the dataset provided.}

\item{data}{a data frame}

\item{y}{a numeric variable}
}
\value{
method used, dataset used, variables used for outliers detected, indices of any detected outliers, scores for the outliers, and values for optional parameters

dataframe with correlations of y variable with all x-variables, p-value, and level of significance for each x variable.
}
\description{
Creates a correlation table using numeric variables.

Creates a correlation table using numeric variables.
}
\examples{
multiOutliers(mtcarsOutliers, method="mahalanobis", alpha=0.1)
multiOutliers(mtcarsOutliers, method="LoF", minPts=5)
multiOutliers(mtcarsOutliers, method="kNN", k=5, threshold=.95)
multiOutliers(mtcarsOutliers, method="iForest", ntrees = 50)
library(MASS)
x <- ryx(Boston, y = "medv")
x
x <- ryx(iris, y = "Sepal.Length")
x
}
