% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ryx.R
\name{print.ryx}
\alias{print.ryx}
\title{Correlation Table}
\usage{
\method{print}{ryx}(x, ...)
}
\arguments{
\item{x}{string, a list of numeric variables. Defaults to all variables in the dataset provided.}

\item{data}{a data frame}

\item{y}{a numeric variable}
}
\value{
dataframe with correlations of y variable with all x-variables, p-value, and level of significance for each x variable.
}
\description{
Creates a correlation table using numeric variables.
}
\examples{
library(MASS)
x <- ryx(Boston, y = "medv")
x
x <- ryx(iris, y = "Sepal.Length")
x
}
