% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ryx}
\alias{print.ryx}
\title{Computes correlation of a dataframe}
\usage{
\method{print}{ryx}(x)
}
\arguments{
\item{x}{an object of class \code{ryx}}
}
\value{
a data frame with 3 components:
\describe{
\item{varNames}{name of the independent variable}
\item{rVal}{coefficient of the correlation between the independent variable and the outcome variable}
\item{pVal}{p-value of the correlation}
\item{sigif}{significance level of the correlation}
}
}
\description{
\code{print.ryx} prints a table of indepndent variables, their correlation with outcome variable,
and the significance of the correlations.
}
\examples{
\dontrun{
x <- ryx(MASS::Boston, y = "medv")
print(x)
}
}
