% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ryx}
\alias{print.ryx}
\title{Print of Correlation for Object Class 'ryx'}
\usage{
\method{print}{ryx}(x, digits = 3)
}
\arguments{
\item{x}{An object of 'ryx' class}

\item{digits}{Rounding digits of the values, default value of 3}
}
\value{
A printing of the correlation table
}
\description{
Given an object of class 'ryx', the \code{print} function would print table with
column names "variable", "r", "p", "sigif". The table also excludes the rownames.
}
\examples{
\dontrun{
#import MASS package for example data
library(MASS)

#create an 'ryx' class object
x <- ryx(Boston, y = "medv")

#print the 'ryx' class object
printR <- print(x)
}

}
