% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ryx.R
\name{print.ryx}
\alias{print.ryx}
\title{Print Method for \code{ryx} Objects}
\usage{
\method{print}{ryx}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ryx}, which is the result of the \code{ryx} function.}
}
\value{
This function prints a table with the following columns:
\itemize{
\item \code{variable}: The name of the independent variable.
\item \code{r}: The correlation coefficient rounded to 3 decimal places.
\item \code{p_display}: The p-value formatted in scientific notation, rounded to 2 decimal places.
\item \code{sigif}: The significance level of the p-value, represented as \code{"***"}, \code{"**"}, \code{"*"}, or \code{" "} based on p-value thresholds.
}
}
\description{
This function prints the results of the \code{ryx} function in a readable and formatted way. It displays the correlation coefficients (\code{r}), p-values, and significance levels (\code{sigif}) for each variable correlated with the dependent variable (\code{y}).
}
\examples{
# Assuming `result` is the result of the `ryx` function:
print(result)
}
