% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxy.R
\name{print.ryx}
\alias{print.ryx}
\title{print.ryx}
\usage{
\method{print}{ryx}(x, digits = 3)
}
\description{
The print function takes correlations
created by the ryx function and prints
a table of correlations that notes when the p value
is < 2e-16
}
\examples{
library(MASS)
x <- ryx(Boston, y="medv")
x

}
