% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ryx.R
\name{print.ryx}
\alias{print.ryx}
\title{Print Function}
\usage{
\method{print}{ryx}(x)
}
\arguments{
\item{x}{an object of class ryx}
}
\value{
data frame
}
\description{
prints out data frame with variables and their correlations,
pvalues, and significance
}
\examples{
library(MASS)
x <- ryx(Boston, y="medv")

}
