% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ryx.R
\name{print.ryx}
\alias{print.ryx}
\title{Print \code{ryx} object}
\usage{
\method{print}{ryx}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{ryx} object}

\item{digits}{(optional) Number of significant digits to use for percentages.}

\item{...}{Optional parameters passed to import functions}
}
\description{
Print the correlations of the given response and explanatory variables.
}
\examples{
x=ryx(mtcars, 'mpg', c('wt', 'hp'))
print(x, digits=4)
}
