% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ryx.R
\name{print.ryx}
\alias{print.ryx}
\title{print.ryx}
\usage{
\method{print}{ryx}(x, digits = 3)
}
\description{
This \code{print.ryx} function can print the ryx objects.

This is wonderful.
}
\examples{
\dontrun{
print.ryx(myryx, digits = 3)
}
}
