% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{Calculate Correlations}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{A data frame containing the variables.}

\item{y}{A string specifying the name of the target variable.}

\item{x}{A character vector of predictor variable names. If not specified, all numeric variables except \code{y} are used.}
}
\value{
An object of class \code{ryx}, containing:
\describe{
\item{y}{The target variable name.}
\item{x}{The predictor variable names.}
\item{df}{A data frame with the correlation results, including variables, correlations, p-values, and significance levels.}
}
}
\description{
This function calculates the correlation of a target variable (\code{y}) with one or more numeric predictor variables (\code{x}).
}
\examples{
library(MASS)
ryx(Boston, y = "medv")

}
