% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{Correlation Table}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{a data frame}

\item{y}{a numeric variable}

\item{x}{string, a list of numeric variables. Defaults to all variables in the dataset provided.}
}
\value{
dataframe with correlations of y variable with all x-variables, p-value, and level of significance for each x variable.
}
\description{
Creates a correlation table using numeric variables.
}
\examples{
ryx(mtcars, "mpg")
ryx(iris, "Sepal.Length")
library(MASS)
ryx(Boston, "medv")
}
