% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryxFunction.R
\name{ryx}
\alias{ryx}
\title{ryx function}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{data frame}

\item{y}{variable used as the outcome variable in correlations}

\item{x}{a vector of independent variables}
}
\value{
a list of y, x, and a data frame reporting correlations of y with each variable in x respectively
}
\description{
This package deals with correlation among variables.

\code{ryx} calculates correlations between one outcome variable and a group
of independent variables, respectively, as well as significance of the correlations.
The final result is reported in the form of a correlation table.
}
\details{
Three generic functions (\code{print.ryx},\code{summary.ryx},\code{plot.ryx}) are
implemented. These functions can be used to report table of ryx class.
}
\examples{
\dontrun{
# create a ryx object
x <- ryx(MASS::Boston, y = medv)

# generic functions
print(x)
summary(x)
plot(x)

# report error message if imput is not class ryx
print.ryx(mtcars)
}

}
