% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{Computes correlation of a data frame.}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{data frame}

\item{y}{variable used as the outcome variable in correlations}

\item{x}{a vector of independent variables}
}
\value{
a data frame with 3 components:
\describe{
\item{y}{name of the outcome variable}
\item{x}{a group of independent variables, each is paired with y to calculate correlation}
\item{df}{data frame for overall correlations}
}
}
\description{
\code{ryx} calculates correlations between specified outcome variable and independent variables.
}
\details{
This function computes correlation between the outcome variable and each independent variable in the specified
group. Each correlation's statistical significance is also computed.
}
\examples{
\dontrun{
x <- ryx(MASS::Boston, y = "medv")
x
print(x)
summary(x)
plot(x)
}
}
\seealso{
\link{print.ryx},\link{summary.ryx},\link{plot.ryx}
}
