% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{Correlation Table of class 'ryx'}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{a dataframe}

\item{y}{a variable}

\item{x}{a vector of variables to create a correlation table, if NULL, return a correlation table of all of the variables
excluding y itself}
}
\value{
A plot with x-axis as the absolute correlation values and y-axis as the variables.
}
\description{
Given a dataframe, a y variable, and a vector x of variables, the \code{ryx} function would
return a list of variable y, vector of variables x, and a dataframe of correlation, p-value, and stars rating of the
p-value.
}
\examples{
\dontrun{
#import MASS package for example data
library(MASS)

#create an 'ryx' class object
x <- ryx(Boston, y = "medv")

}


}
