% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{Conduct Correlation Analysis}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{A data frame}

\item{y}{The dependent variable}

\item{x}{The independent variables}
}
\value{
A list of class \code{ryx} containing:
\itemize{
\item \code{y} The dependent variable
\item \code{x} The independent variables
\item \code{df} A data frame with columns
\item \code{r} correlation coefficient)
\item \code{p} (p-value)
\item \code{sigif} (significance)
}
}
\description{
This is a function that conducts correlation tests between one dependent variable and one or more independent variables.
If the explanatory variable is not specified, it will conduct correlation analysis on all numeric variables in the dataset.
The function returns a list of elements, including correlation coefficients, p-values, and the significance levels.
}
\examples{
x <- ryx(Boston, y = "medv")
print(x)  # Print results of the correlation analysis
summary(x)  # Get a summary of the results
plot(x)  # Visualize the correlations
}
