% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxy.R
\name{ryx}
\alias{ryx}
\title{ryx}
\usage{
ryx(data, y, x)
}
\description{
The ryx function prints a
table of correlations between a dependent variable y and one
or more independent x variables.
}
\examples{
ryx(mtcars, "mpg")

}
