% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{ryx Function}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{A data frame.}

\item{y}{The dependent variable.}

\item{x}{A vector of independent variables.}
}
\value{
A summary of correlations, p-values, and significance.
}
\description{
function that takes in data, y variable and x variables
and describes  correlations, p-values, and significance
of the y variable against the x variables
}
\examples{

ryx(mtcars, y="mpg")

}
