% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment4.R
\name{ryx}
\alias{ryx}
\title{Test correlations}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{the dataset to test on}

\item{y}{the name of the quantitative response variable (str)}

\item{x}{a vector of names of quantitative explanatory variables. If not provided, then automatically use all the quantitative variables.}

\item{...}{optional parameters passed to import functions}
}
\value{
\describe{
\item{y}{The name of response variable}
\item{x}{A vector of names of explanatory variables}
\item{df}{A dataframe of correlations}
}
}
\description{
Test the correlations between the response variable and all the other explanatory variables provided.
}
\examples{
# x is missing
ryx(mtcars, 'mpg')

# give x
ryx(mtcars, 'mpg', c('hp', 'wt'))
}
