% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx.R
\name{ryx}
\alias{ryx}
\title{ryx}
\usage{
ryx(data, y, x)
}
\value{
ryx
}
\description{
This \code{ryx} function can create a ryx type.

This is wonderful.
}
\examples{
\dontrun{
library(MASS)
myryx <- ryx(Boston, y="medv")
}
}
