% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx_methods.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Summary method for ryx objects}
\usage{
\method{summary}{ryx}(object, ...)
}
\arguments{
\item{object}{An object of class ryx.}
}
\description{
Summary method for ryx objects
}
\examples{
library(MASS)
x <- ryx(Boston, y = "medv")
summary(x)

}
