% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Correlation Table}
\usage{
\method{summary}{ryx}(x, ..)
}
\arguments{
\item{x}{string, a list of numeric variables. Defaults to all variables in the dataset provided.}

\item{data}{a data frame}

\item{y}{a numeric variable}
}
\value{
method used, dataset used, variables used for outliers detected, indices of any detected outliers, scores for the outliers, and values for optional parameters
}
\description{
Creates a correlation table using numeric variables.
}
\examples{
library(MASS)
x <- ryx(Boston, y = "medv")
summary(x)
x <- ryx(iris, y = "Sepal.Length")
summary(x)
}
