% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Correlation Table}
\usage{
\method{summary}{ryx}(x, ..)
}
\arguments{
\item{x}{object inherited from ryx function}

\item{data}{a data frame}

\item{...}{(optional), any additional parameters}
}
\value{
summary information for ryx object
}
\description{
Creates a correlation table using numeric variables.
}
\examples{
library(MASS)
x <- ryx(Boston, y = "medv")
summary(x)
x <- ryx(iris, y = "Sepal.Length")
summary(x)
}
