% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Computes correlation of a data frame}
\usage{
\method{summary}{ryx}(x)
}
\arguments{
\item{x}{an object of class \code{ryx}}
}
\description{
\code{summary.ryx} print a text message summarizing the independent variables, the range of correlation coefficient,
and the number of significantly correlated variables.
}
\examples{
\dontrun{
x <- ryx(MASS::Boston, y = "medv")
summary(x)
}
}
