% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Summary of Correlation for Object Class 'ryx'}
\usage{
\method{summary}{ryx}(x, digits = 3)
}
\arguments{
\item{x}{An object of 'ryx' class}

\item{digits}{Digits for rounding of the values, Default value is at 3}
}
\value{
A summary statement of the median correlation value, range of correlation values, and amount
of variables with p-value < 0.05
}
\description{
Given an object of class 'ryx', the \code{summary} function would print out the
description of the 'ryx' object class. The description includes the variables that are compared to each other,
the median absolute correlation value, the range of correlation values, and amount of variables with
p values less than 0.05.
}
\examples{
\dontrun{
#import MASS package for example data
library(MASS)

#create an 'ryx' class object
x <- ryx(Boston, y = "medv")

#summary of the 'ryx' class object
summaryR <- summary(x)
}

}
