% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Summary function for \code{ryx} Objects}
\usage{
\method{summary}{ryx}(object)
}
\arguments{
\item{object}{An object of class \code{ryx}, which is the result of the \code{ryx} function.}
}
\value{
This function returns \code{NULL} but prints the following summary to the console:
\itemize{
\item The median absolute correlation value.
\item The range of correlation values (minimum and maximum).
\item The number of significant correlations (p-value < 0.05) out of the total number of correlations.
}
}
\description{
This function provides a summary of the correlation results stored in a \code{ryx} object. It calculates and prints key statistics such as the median absolute correlation, the range of correlations, and the number of significant correlations (p-value < 0.05).
}
\examples{
x <- ryx(Boston, y = "medv")
print(x)
summary(x) 
}
