% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxy.R
\name{summary.ryx}
\alias{summary.ryx}
\title{summary.ryx}
\usage{
\method{summary}{ryx}(x)
}
\description{
The summary function takes correlations
created by the ryx function and prints
a summary of the median correlation, the range,
and how many variables are significant
at the p < 0.05 level.
}
\examples{
library(MASS)
x <- ryx(Boston, y="medv")
summary(x)
}
