% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Summary Function}
\usage{
\method{summary}{ryx}(ryx)
}
\arguments{
\item{x}{an object of class ryx}
}
\value{
3 lines of text
}
\description{
summarizes which variables are included, the median
and range of correlations, and the number of significant variables
}
\examples{
x <- ryx(mtcars, y="mpg")
summary(x)

}
