% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{Summary \code{ryx} object}
\usage{
\method{summary}{ryx}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{ryx} object}

\item{digits}{(optional) Number of significant digits to use for percentages.}

\item{...}{Optional parameters passed to import functions}
}
\description{
Summary the correlations of the given response and explanatory variables.
}
\examples{
x=ryx(mtcars, 'mpg', c('wt', 'hp'))
summary(x, digits=4)
}
