% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ryx.R
\name{summary.ryx}
\alias{summary.ryx}
\title{summary.ryx}
\usage{
\method{summary}{ryx}(x, digits = 3)
}
\description{
This \code{summary.ryx} function can summarize the ryx objects for you.

This is wonderful.
}
